+++ src/common/profile.ui.ts
@@ -3803,6 +3803,11 @@ ipcRenderer.on('profile-item-added', (event, type, profile, item) => {
     $('#btn-refresh').addClass('btn-warning');
 });
 
+ipcRenderer.on('profile-item-updated', (event, type, profile, idx, item) => {
+    filesChanged = true;
+    $('#btn-refresh').addClass('btn-warning');
+});
+
 ipcRenderer.on('profile-item-removed', (event, type, profile, idx) => {
     filesChanged = true;
     $('#btn-refresh').addClass('btn-warning');